AgilentLaserCombiner: Prevent VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10919 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/AgilentLaserCombiner/AgilentLaserCombiner.cpp
@@ -571,7 +571,7 @@ int LCShutter::OnState(MM::PropertyBase *pProp, MM::ActionType eAct)
          os >> seq;
          load[i] = (unsigned char) seq;
       }
-      int ret = LaserBoardSetSequenceState(sequence.size(), load);
+      int ret = LaserBoardSetSequenceState(static_cast<unsigned int>(sequence.size()), load);
       delete(load);
       return ret;
    }
